requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.1.0

# MuJoCo
mujoco>=3.0.0
gymnasium[mujoco]

# Experiment Tracking
wandb>=0.16.0

# Explainability/Interpretability
captum>=0.6.0
shap>=0.43.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
tqdm>=4.65.0
pandas>=2.0.0

# LLM & Agents
anthropic>=0.39.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Vector Store & Embeddings (for RAG)
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Async utilities
aiohttp>=3.9.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# Development
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
